# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NPTC Catalogue Maintenance Platform",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/v1/catalogue/entries/{business_key}/bindings": {
      "get": {
        "tags": [
          "catalogue"
        ],
        "summary": "An entry's SNOMED CT code bindings, including retired ones",
        "description": "Retired bindings are included (FR-08): a client holding a code that\nhas been inactivated learns so here, together with the reason and any\nsuccessor code.",
        "operationId": "read_bindings_api_v1_catalogue_entries__business_key__bindings_get",
        "parameters": [
          {
            "name": "business_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^NPTC-([0-9]{6,})$",
              "description": "The entry's public identifier, e.g. `NPTC-000247` (FR-03).",
              "examples": [
                "NPTC-000247"
              ],
              "title": "Business Key"
            },
            "description": "The entry's public identifier, e.g. `NPTC-000247` (FR-03)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BindingList"
                }
              }
            }
          },
          "401": {
            "description": "A credential was presented and could not be verified. Sending no credential at all is not an error on these endpoints.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No entry in the published catalogue has this business key. An entry that exists but is not published (draft, deprecated or withdrawn) is reported identically, on purpose - a distinguishable response would confirm the key exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "A query or path parameter was unprocessable - a business key that is not `NPTC-nnnnnn`, a blank search query, a cursor this API did not issue (including one issued for a different `q` or a different filter set), a `limit` outside its range, or a `filter.*` parameter naming a facet this endpoint does not offer, an operator the facet does not support, or a value the property cannot hold. A filter is never silently ignored.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "catalogue-admin"
        ],
        "summary": "Bind a SNOMED CT code to a catalogue entry",
        "operationId": "bind_code_api_v1_catalogue_entries__business_key__bindings_post",
        "parameters": [
          {
            "name": "business_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^NPTC-([0-9]{6,})$",
              "description": "The entry's public identifier, e.g. `NPTC-000247` (FR-03).",
              "examples": [
                "NPTC-000247"
              ],
              "title": "Business Key"
            },
            "description": "The entry's public identifier, e.g. `NPTC-000247` (FR-03)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BindCodeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BindingWriteResult"
                }
              }
            },
            "headers": {
              "Location": {
                "description": "The entry (`GET {business_key}`) the new binding was added to.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "No credential, or one that could not be verified.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The caller is authenticated but does not hold `catalogue.edit_published`, or holds it but has not completed the MFA step-up this permission requires (the response then also carries a `WWW-Authenticate` step-up challenge).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No catalogue entry, or no active code binding, matches the given identifier.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The request is well-formed but conflicts with the current state of the system - a second active binding on this entry, a successor code already actively bound elsewhere (including two concurrent requests racing for the same entry or code), `/replacement`'s successor naming the same code it is meant to replace, or a stale `expected_row_version` (FR-38) because someone else changed this entry since it was loaded - that variant carries `business_key`, `expected_row_version`, `current_row_version`, `conflicts[]` and `changed_by`/`changed_at`, so the caller can reconcile rather than retry blind. A code already retired, or with no binding at all, is a 404 here rather than a 409: every route below addresses a binding by its currently-*active* code, so a retired one is simply not addressable this way any more, not a conflicting state.",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/VersionConflictResponse"
                    }
                  ],
                  "title": "Response 409 Bind Code Api V1 Catalogue Entries  Business Key  Bindings Post"
                }
              }
            }
          },
          "422": {
            "description": "A field failed validation - a malformed or Verhoeff-failing SCTID, an unrecognised edition hint, a blank `fsn`/`au_preferred_term`, or a changelog note that does not meet FR-37.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "A platform-side invariant failed, not a caller mistake - e.g. re-reading a binding this same request just wrote could not find it. Not produced by anything a well-formed request can trigger on its own; retrying will not clear it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/catalogue/entries/{business_key}/bindings/{code}/retirement": {
      "post": {
        "tags": [
          "catalogue-admin"
        ],
        "summary": "Retire an entry's active code binding",
        "operationId": "retire_binding_api_v1_catalogue_entries__business_key__bindings__code__retirement_post",
        "parameters": [
          {
            "name": "business_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^NPTC-([0-9]{6,})$",
              "description": "The entry's public identifier, e.g. `NPTC-000247` (FR-03).",
              "examples": [
                "NPTC-000247"
              ],
              "title": "Business Key"
            },
            "description": "The entry's public identifier, e.g. `NPTC-000247` (FR-03)."
          },
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetireBindingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BindingWriteResult"
                }
              }
            }
          },
          "401": {
            "description": "No credential, or one that could not be verified.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "The caller is authenticated but does not hold `catalogue.edit_published`, or holds it but has not completed the MFA step-up this permission requires (the response then also carries a `WWW-Authenticate` step-up challenge).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No catalogue entry, or no active code binding, matches the given identifier.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The request is well-formed but conflicts with the current state of the system - a second active binding on this entry, a successor code already actively bound elsewhere (including two concurrent requests racing for the same entry or code), `/replacement`'s successor naming the same code it is meant to replace, or a stale `expected_row_version` (FR-38) because someone else changed this entry since it was loaded - that variant carries `business_key`, `expected_row_version`, `current_row_version`, `conflicts[]` and `changed_by`/`changed_at`, so the caller can reconcile rather than retry blind. A code already retired, or with no binding at all, is a 404 here rather than a 409: every route below addresses a binding by its currently-*active* code, so a retired one is simply not addressable this way any more, not a conflicting state.",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    },
                    {
                      "$ref": "#/components/schemas/VersionConflictResponse"
                    }
                  ],
                  "title": "Response 409 Retire Binding Api V1 Catalogue Entries  Business Key  Bindings  Code  Retirement Post"
                }
              }
            }
          },
          "422": {
            "description": "A field failed validation - a malformed or Verhoeff-failing SCTID, an unrecognised edition hint, a blank `fsn`/`au_preferred_term`, or a changelog note that does not meet FR-37.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "A platform-side invariant failed, not a caller mistake - e.g. re-reading a binding this same request just wrote could not find it. Not produced by anything a well-formed request can trigger on its own; retrying will not clear it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BindCodeRequest": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "fsn": {
            "type": "string",
            "minLength": 1,
            "title": "Fsn"
          },
          "au_preferred_term": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Au Preferred Term"
          },
          "edition_hint": {
            "$ref": "#/components/schemas/CodeBindingEditionHint",
            "default": "unknown"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "expected_row_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Row Version"
          }
        },
        "type": "object",
        "required": [
          "code",
          "fsn",
          "reason",
          "expected_row_version"
        ],
        "title": "BindCodeRequest",
        "description": "The body of `POST /catalogue/entries/{business_key}/bindings`.\n\n`code` is a string end-to-end (FR-06) - `nptc.catalogue.bindings.\ncreate_binding` validates it via `nptc_shared.sctid.SCTID` before any\nrow is touched. `fsn`/`au_preferred_term` are carried through exactly\nas submitted (FR-82); this screen is the one place they are meant to be\ntyped in, and neither is cleaned or re-derived here or anywhere else."
      },
...
      "BindingList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Binding"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "BindingList"
      },
...
      "BindingWriteResult": {
        "properties": {
          "binding": {
            "$ref": "#/components/schemas/Binding"
          },
          "row_version": {
            "type": "integer",
            "title": "Row Version"
          }
        },
        "type": "object",
        "required": [
          "binding",
          "row_version"
        ],
        "title": "BindingWriteResult",
        "description": "`bind_code`/`retire_binding`'s response: the affected binding, plus\nthe entry's new `row_version` - mirroring `catalogue_properties.\nPropertyValuesWriteResult`, so a client never has to re-fetch the entry\njust to learn its next lock token. Declared here, not in\n`catalogue_shared.py`: that module's `Binding`/`BindingList` are shared\nwith the *public* read route (`catalogue.py`), and widening them with\nan admin-only `row_version` field would break\n`test_api_public_response_hygiene.py`."
      },
...
      "CodeBindingEditionHint": {
        "type": "string",
        "enum": [
          "au",
          "int",
          "unknown"
        ],
        "title": "CodeBindingEditionHint"
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "Every refusal this API makes has this shape - one sentence saying\nwhat to do next, and deliberately nothing else. It never names a role,\na permission or an internal identifier (FR-44, NFR-04)."
      },
...
      "RetireBindingRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "expected_row_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Row Version"
          }
        },
        "type": "object",
        "required": [
          "reason",
          "expected_row_version"
        ],
        "title": "RetireBindingRequest"
      },
...
    }
  }
}